Add Simulation class
At the moment this class puts all of the passengers into an ArrayList, from which it randomly removes passengers and adds them to the queue. The code loops until the for 15000 milliseconds, testing the condition rand.nextInt(1000) == 1 to simulate the chance of a passenger turning up. Finally it pops them all to test that it worked. The unneeded print statements and popAndPrintQueue method can easily be removed.